move folders fix
